Refactor View to ViewModel
